+++ CHANGELOG.md
@@ -24,6 +24,7 @@
   subclasses. This involved moving functionality from `ReactArezComponent` back to `arez.Component` class,
   marking the `shouldUpdateOnChange()` as package-access and final and updating the annotation processor to
   generate the native glue code only if there is any logic generated to support `shouldUpdateOnChange`.
+* **\[core\]** Remove the mechanisms by which component authors can update react state.
 
 ### [v0.101](https://github.com/react4j/react4j/tree/v0.101) (2018-10-09)
 [Full Changelog](https://github.com/react4j/react4j/compare/v0.100...v0.101)